Remove DesktopBuildServerController
The type `BuildServerController` only had a single derived type. This
was likely done back in the netstandard1.3 days when the APIs weren't
sufficient to share code between the desktop and coreclr builds. Now
though there is no reason for this abstraction and it just makes the
code harder to read.
