+++ src/Compilers/Server/VBCSCompiler/BuildServerController.cs
@@ -10,6 +10,7 @@ using System.Threading.Tasks;
 using System.Globalization;
 using Microsoft.CodeAnalysis.CommandLine;
 using System.Runtime.InteropServices;
+using System.Collections.Specialized;
 
 namespace Microsoft.CodeAnalysis.CompilerServer
 {
@@ -17,8 +18,17 @@ namespace Microsoft.CodeAnalysis.CompilerServer
     /// Base type for the build server code.  Contains the basic logic for running the actual server, startup 
     /// and shutdown.
     /// </summary>
-    internal abstract class BuildServerController
+    internal sealed class BuildServerController
     {
+        internal const string KeepAliveSettingName = "keepalive";
+
+        private readonly NameValueCollection _appSettings;
+
+        internal BuildServerController(NameValueCollection appSettings)
+        {
+            _appSettings = appSettings;
+        }
+
         internal int Run(string[] args)
         {
             string pipeName;
@@ -39,30 +49,78 @@ namespace Microsoft.CodeAnalysis.CompilerServer
                 : RunServer(pipeName, tempPath, cancellationToken: cancellationTokenSource.Token);
         }
 
-        protected internal abstract TimeSpan? GetKeepAliveTimeout();
-
-        protected abstract string GetDefaultPipeName();
-
-        protected abstract IClientConnectionHost CreateClientConnectionHost(string pipeName);
-
-        protected abstract Task<Stream> ConnectForShutdownAsync(string pipeName, int timeout);
+        internal TimeSpan? GetKeepAliveTimeout()
+        {
+            try
+            {
+                int keepAliveValue;
+                string keepAliveStr = _appSettings[KeepAliveSettingName];
+                if (int.TryParse(keepAliveStr, NumberStyles.Integer, CultureInfo.InvariantCulture, out keepAliveValue) &&
+                    keepAliveValue >= 0)
+                {
+                    if (keepAliveValue == 0)
+                    {
+                        // This is a one time server entry.
+                        return null;
+                    }
+                    else
+                    {
+                        return TimeSpan.FromSeconds(keepAliveValue);
+                    }
+                }
+                else
+                {
+                    return ServerDispatcher.DefaultServerKeepAlive;
+                }
+            }
+            catch (Exception e)
+            {
+                CompilerServerLogger.LogException(e, "Could not read AppSettings");
+                return ServerDispatcher.DefaultServerKeepAlive;
+            }
+        }
 
         /// <summary>
         /// Was a server running with the specified session key during the execution of this call?
         /// </summary>
-        protected virtual bool? WasServerRunning(string pipeName)
+        private bool? WasServerRunning(string pipeName)
         {
-            return null;
+            string mutexName = BuildServerConnection.GetServerMutexName(pipeName);
+            return BuildServerConnection.WasServerMutexOpen(mutexName);
         }
 
-        protected virtual int RunServerCore(string pipeName, IClientConnectionHost connectionHost, IDiagnosticListener listener, TimeSpan? keepAlive, CancellationToken cancellationToken)
+        private IClientConnectionHost CreateClientConnectionHost(string pipeName)
         {
-            CompilerServerLogger.Log("Keep alive timeout is: {0} milliseconds.", keepAlive?.TotalMilliseconds ?? 0);
-            FatalError.Handler = FailFast.OnFatalException;
+            var compilerServerHost = CreateCompilerServerHost();
+            return CreateClientConnectionHostForServerHost(compilerServerHost, pipeName);
+        }
 
-            var dispatcher = new ServerDispatcher(connectionHost, listener);
-            dispatcher.ListenAndDispatchConnections(keepAlive, cancellationToken);
-            return CommonCompiler.Succeeded;
+        internal static ICompilerServerHost CreateCompilerServerHost()
+        {
+            // VBCSCompiler is installed in the same directory as csc.exe and vbc.exe which is also the 
+            // location of the response files.
+            var clientDirectory = AppDomain.CurrentDomain.BaseDirectory;
+            var sdkDirectory = BuildClient.GetSystemSdkDirectory();
+
+            return new DesktopCompilerServerHost(clientDirectory, sdkDirectory);
+        }
+
+        internal static IClientConnectionHost CreateClientConnectionHostForServerHost(
+            ICompilerServerHost compilerServerHost,
+            string pipeName)
+        {
+            return new NamedPipeClientConnectionHost(compilerServerHost, pipeName);
+        }
+
+        private async Task<Stream> ConnectForShutdownAsync(string pipeName, int timeout)
+        {
+            return await BuildServerConnection.TryConnectToServerAsync(pipeName, timeout, cancellationToken: default).ConfigureAwait(false);
+        }
+
+        private string GetDefaultPipeName()
+        {
+            var clientDirectory = AppDomain.CurrentDomain.BaseDirectory;
+            return BuildServerConnection.GetPipeNameForPathOpt(clientDirectory);
         }
 
         internal int RunServer(
@@ -71,15 +129,49 @@ namespace Microsoft.CodeAnalysis.CompilerServer
             IClientConnectionHost clientConnectionHost = null,
             IDiagnosticListener listener = null,
             TimeSpan? keepAlive = null,
-            CancellationToken cancellationToken = default(CancellationToken))
+            CancellationToken cancellationToken = default)
         {
             keepAlive = keepAlive ?? GetKeepAliveTimeout();
             listener = listener ?? new EmptyDiagnosticListener();
             clientConnectionHost = clientConnectionHost ?? CreateClientConnectionHost(pipeName);
-            return RunServerCore(pipeName, clientConnectionHost, listener, keepAlive, cancellationToken);
+
+            // Grab the server mutex to prevent multiple servers from starting with the same
+            // pipename and consuming excess resources. If someone else holds the mutex
+            // exit immediately with a non-zero exit code
+            var mutexName = BuildServerConnection.GetServerMutexName(pipeName);
+            bool createdNew;
+            using (var serverMutex = BuildServerConnection.OpenOrCreateMutex(name: mutexName,
+                                                                             createdNew: out createdNew))
+            {
+                if (!createdNew)
+                {
+                    return CommonCompiler.Failed;
+                }
+
+                CompilerServerLogger.Log("Keep alive timeout is: {0} milliseconds.", keepAlive?.TotalMilliseconds ?? 0);
+                FatalError.Handler = FailFast.OnFatalException;
+
+                var dispatcher = new ServerDispatcher(clientConnectionHost, listener);
+                dispatcher.ListenAndDispatchConnections(keepAlive, cancellationToken);
+                return CommonCompiler.Succeeded;
+            }
         }
 
-        internal int RunShutdown(string pipeName, bool waitForProcess = true, TimeSpan? timeout = null, CancellationToken cancellationToken = default(CancellationToken))
+        internal static int CreateAndRunServer(
+            string pipeName,
+            string tempPath,
+            IClientConnectionHost clientConnectionHost = null,
+            IDiagnosticListener listener = null,
+            TimeSpan? keepAlive = null,
+            NameValueCollection appSettings = null,
+            CancellationToken cancellationToken = default)
+        {
+            appSettings ??= new NameValueCollection();
+            var controller = new BuildServerController(appSettings);
+            return controller.RunServer(pipeName, tempPath, clientConnectionHost, listener, keepAlive, cancellationToken);
+        }
+
+        internal int RunShutdown(string pipeName, bool waitForProcess = true, TimeSpan? timeout = null, CancellationToken cancellationToken = default)
         {
             return RunShutdownAsync(pipeName, waitForProcess, timeout, cancellationToken).GetAwaiter().GetResult();
         }
@@ -91,7 +183,7 @@ namespace Microsoft.CodeAnalysis.CompilerServer
         /// This function will return success if at any time in the function the server is determined to no longer
         /// be running.
         /// </summary>
-        internal async Task<int> RunShutdownAsync(string pipeName, bool waitForProcess = true, TimeSpan? timeout = null, CancellationToken cancellationToken = default(CancellationToken))
+        internal async Task<int> RunShutdownAsync(string pipeName, bool waitForProcess = true, TimeSpan? timeout = null, CancellationToken cancellationToken = default)
         {
             if (WasServerRunning(pipeName) == false)
             {
