+++ src/Compilers/Server/VBCSCompilerTests/DesktopBuildServerControllerTests.cs
@@ -6,29 +6,29 @@ using Xunit;
 
 namespace Microsoft.CodeAnalysis.CompilerServer.UnitTests
 {
-    public sealed class DesktopBuildServerControllerTests
+    public sealed class BuildServerControllerTests
     {
         public sealed class GetKeepAliveTimeoutTests
         {
             private readonly NameValueCollection _appSettings = new NameValueCollection();
-            private readonly DesktopBuildServerController _controller;
+            private readonly BuildServerController _controller;
 
             public GetKeepAliveTimeoutTests()
             {
-                _controller = new DesktopBuildServerController(_appSettings);
+                _controller = new BuildServerController(_appSettings);
             }
 
             [Fact]
             public void Simple()
             {
-                _appSettings[DesktopBuildServerController.KeepAliveSettingName] = "42";
+                _appSettings[BuildServerController.KeepAliveSettingName] = "42";
                 Assert.Equal(TimeSpan.FromSeconds(42), _controller.GetKeepAliveTimeout());
             }
 
             [Fact]
             public void InvalidNumber()
             {
-                _appSettings[DesktopBuildServerController.KeepAliveSettingName] = "dog";
+                _appSettings[BuildServerController.KeepAliveSettingName] = "dog";
                 Assert.Equal(ServerDispatcher.DefaultServerKeepAlive, _controller.GetKeepAliveTimeout());
             }
